octoshop/requirements.txt
amqp==2.6.1
asgiref==3.3.1
astroid==2.5.2
billiard==3.6.3.0
cairocffi==1.2.0
CairoSVG==2.5.1
celery==4.4.2
cffi==1.14.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.5
django-debug-toolbar==3.2
django-extensions==3.1.0
django-parler==2.2
html5lib==1.1
isort==5.8.0
kombu==4.6.11
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.1.0
prompt-toolkit==3.0.10
pycparser==2.20
pydot==1.4.1
pylint==2.7.4
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyparsing==2.4.7
Pyphen==0.10.0
python-memcached==1.59
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
toml==0.10.2
vine==1.3.0
wcwidth==0.2.5
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
